@@ -2,15 +2,15 @@ from py_air_control_exporter import app, metrics
 
 
 def test_metrics(mock_fetcher):
-    """metrics endpoint produces the expected metrics"""
+    """Metrics endpoint produces the expected metrics"""
     mock_fetcher.return_value = metrics.AirControlStatus(
         status=metrics.Status(fan_speed=0, iaql=1, is_manual=True, is_on=True, pm25=2),
         filters=metrics.Filters(
             filters={
-                "0": metrics.Filter(hours=0, type=""),
-                "1": metrics.Filter(hours=185, type="A3"),
-                "2": metrics.Filter(hours=2228, type="C7"),
-            }
+                "0": metrics.Filter(hours=0, filter_type=""),
+                "1": metrics.Filter(hours=185, filter_type="A3"),
+                "2": metrics.Filter(hours=2228, filter_type="C7"),
+            },
         ),
     )
     response = app.create_app(mock_fetcher).test_client().get("/metrics")
@@ -26,7 +26,7 @@ def test_metrics(mock_fetcher):
 
 
 def test_metrics_failure(mock_fetcher):
-    """metrics endpoint should produce a sampling error counter on error"""
+    """Metrics endpoint should produce a sampling error counter on error"""
     mock_fetcher.side_effect = Exception()
     test_client = app.create_app(mock_fetcher).test_client()
     response = test_client.get("/metrics")
@@ -36,7 +36,7 @@ def test_metrics_failure(mock_fetcher):
 
 
 def test_metrics_fetched_again(mock_fetcher):
-    """check that status is fetched every time metrics are pulled"""
+    """Check that status is fetched every time metrics are pulled"""
     assert mock_fetcher.call_count == 0
     test_client = app.create_app(mock_fetcher).test_client()
     assert mock_fetcher.call_count == 1
